add code block from gfg
This commit represents the code which scraps the code block content from
the GFG page.

GFG_URL = `https://www.geeksforgeeks.org/python-basics/`

This code only gets the code block from the multiple blocks on the page.
Using a function to get the data from the website once, then using the
local file to work with. There is possible to get blocked, if there are
multiple calls in a period of time. Using Beautiful Soup, looking for
elements and then from list of each element, getting the text out of it.